requirements.txt
fastapi
uvicorn
transformers
datasets
beautifulsoup4
requests
torch
faiss-gpu
sentence_transformers